FSDP: Mistral(mathstral) sbatch file - MISTRAL MODEL SUPPORT (#385)
* Rename 2.distributed-training-mistral.sbatch to 2.distributed-training-mixtral.sbatch

renaming to mixtral because the args here/ referenced in model _type flag are specific to mixtral models

* commit mistral-mathstral-sbatch

contains updated sbatch for mistral models. Before these changes, only mixtral was supported. This PR is part of other PRs that Armando and I will incorporate. This has been tested on hyperpod already.

* Update 3.distributed-training-mistral-mathstral.sbatch

* Update 3.distributed-training-mistral-mathstral.sbatch

* Update README.md

* Update README.md